+++ apicast/src/threescale_utils.lua
@@ -179,6 +179,14 @@ function _M.release_redis(red)
   red:set_keepalive(redis_conf.keepalive, redis_conf.poolsize)
 end
 
+local xml_header_len = string.len('<?xml version="1.0" encoding="UTF-8"?>')
+
+function _M.match_xml_element(xml, element, value)
+  if not xml then return nil end
+  local pattern = string.format('<%s>%s</%s>', element, value, element)
+  return string.find(xml, pattern, xml_header_len, xml_header_len, true)
+end
+
 -- error and exist
 function _M.error(...)
   if ngx.get_phase() == 'timer' then