+++ apicast/src/configuration_loader/remote_v2.lua
@@ -12,6 +12,7 @@ local user_agent = require 'user_agent'
 local cjson = require 'cjson'
 local resty_env = require 'resty.env'
 local re = require 'ngx.re'
+local configuration = require 'configuration'
 
 local _M = {
   _VERSION = '0.1'
@@ -69,6 +70,10 @@ local function status_code_error(response)
   }, status_error_mt)
 end
 
+local function array()
+  return setmetatable({}, cjson.empty_array_mt)
+end
+
 function _M:index(host)
   local http_client = self.http_client
 
@@ -88,7 +93,7 @@ function _M:index(host)
     return nil, 'missing environment'
   end
 
-  local url = resty_url.join(self.endpoint, '/', env, '.json?', encode_args({ host = host }))
+  local url = resty_url.join(self.endpoint, env .. '.json?' .. encode_args({ host = host }))
   local res, err = http_client.get(url)
 
   if not res and err then
@@ -101,15 +106,22 @@ function _M:index(host)
   if res.status == 200 then
     local json = cjson.decode(res.body)
 
-    local configs = {}
+    local config = { services = array(), oidc = array() }
 
     local proxy_configs = json.proxy_configs or {}
 
     for i=1, #proxy_configs do
-      configs[i] = proxy_configs[i].proxy_config.content
+      local proxy_config = proxy_configs[i].proxy_config
+      local service = configuration.parse_service(proxy_config.content)
+      local issuer, oidc_config = self:oidc_issuer_configuration(service)
+
+      if issuer then
+        config.oidc[i] = { issuer = issuer, config = oidc_config }
+      end
+      config.services[i] = proxy_config.content
     end
 
-    return cjson.encode({ services = configs })
+    return cjson.encode(config)
   else
     return nil, 'invalid status'
   end
@@ -138,7 +150,8 @@ function _M:call(environment)
   if not env then
     return nil, 'missing environment'
   end
-  local configs = {}
+
+  local configs = { services = array(), oidc = array() }
 
   local res, err = self:services()
 
@@ -158,11 +171,13 @@ function _M:call(environment)
     end
   end
 
-  for i, c in ipairs(configs) do
-    configs[i] = c.content
+  for i=1, #configs do
+    configs.services[i] = configs[i].content
+    configs.oidc[i] = configs[i].oidc
+    configs[i] = nil
   end
 
-  return cjson.encode({ services = configs })
+  return cjson.encode(configs)
 end
 
 local services_subset = function()
@@ -203,10 +218,61 @@ function _M:services()
   if res.status == 200 then
     local json = cjson.decode(res.body)
 
-    return json.services or {}
+    return json.services or array()
   else
     return nil, status_code_error(res)
   end
+end
+
+local function openid_configuration_url(endpoint)
+  return resty_url.join(endpoint, '.well-known/openid-configuration')
+end
+
+function _M:oidc_issuer_configuration(service)
+  local http_client = self.http_client
+
+  if not http_client then
+    return nil, 'not initialized'
+  end
+
+  local endpoint = service.oidc.issuer_endpoint
+
+  if not endpoint then
+    return nil, 'no OIDC endpoint'
+  end
+
+  local uri = openid_configuration_url(endpoint)
+  local res = http_client.get(uri)
+
+  if res.status ~= 200 then
+    ngx.log(ngx.ERR, 'failed to get OIDC Provider from ', uri, ' status: ', res.status, ' body: ', res.body)
+    return nil, 'could not get OpenID Connect configuration'
+  end
+
+  local config = res.headers.content_type == 'application/json' and cjson.decode(res.body)
+
+  if not config then
+    ngx.log(ngx.ERR, 'invalid OIDC Provider, expected application/json got:  ', res.headers.content_type, ' body: ', res.body)
+    return nil, 'invalid JSON'
+  end
+
+  res = http_client.get(config.issuer)
+
+  if res.status ~= 200 then
+    ngx.log(ngx.ERR, 'failed to get OIDC Issuer from ', uri, ' status: ', res.status, ' body: ', res.body)
+    return nil, 'could not get OpenID Connect Issuer'
+  end
+
+  local issuer = res.headers.content_type == 'application/json' and cjson.decode(res.body)
+
+  if not issuer then
+    ngx.log(ngx.ERR, 'invalid OIDC Issuer, expected application/json got:  ', res.headers.content_type, ' body: ', res.body)
+    return nil, 'invalid JSON'
+  end
+
+  issuer.openid = config
+
+  return config.issuer, issuer
 end
 
 function _M:config(service, environment, version)
@@ -231,16 +297,23 @@ function _M:config(service, environment, version)
   local res, err = http_client.get(url)
 
   if not res and err then
-    ngx.log(ngx.DEBUG, 'services get error: ', err, ' url: ', url)
+    ngx.log(ngx.ERR, 'services get error: ', err, ' url: ', url)
     return nil, err
   end
 
   ngx.log(ngx.DEBUG, 'services get status: ', res.status, ' url: ', url, ' body: ', res.body)
 
   if res.status == 200 then
-    local json = cjson.decode(res.body)
+    local proxy_config = cjson.decode(res.body).proxy_config
 
-    return json.proxy_config
+    local config_service = configuration.parse_service(proxy_config.content)
+    local issuer, oidc_config = self:oidc_issuer_configuration(config_service)
+
+    if issuer then
+      proxy_config.oidc = { issuer = issuer, config = oidc_config }
+    end
+
+    return proxy_config
   else
     return nil, status_code_error(res)
   end