+++ CHANGELOG.md
@@ -11,11 +11,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - APIcast module `balancer` method now accepts optional balancer [PR #362](https://github.com/3scale/apicast/pull/362)
 - Extracted lua-resty-url [PR #384](https://github.com/3scale/apicast/pull/384)
 - Extracted lua-resty-env [PR #386](https://github.com/3scale/apicast/pull/386)
-- Do not load all services when APICAST_SERVICES is set [PR #388](https://github.com/3scale/apicast/pull/388)
+- Do not load all services when APICAST\_SERVICES is set [PR #388](https://github.com/3scale/apicast/pull/388)
 
 ### Added
 
 - APIcast published to [luarocks.org](https://luarocks.org/modules/3scale/apicast) [PR #366](https://github.com/3scale/apicast/pull/366)
+- Support for passing remote configuratio URL through the CLI [PR #389](https://github.com/3scale/apicast/pull/389)
 
 ## [3.1.0-alpha1] - 2017-05-05