+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [3.1.0-beta1] - 2017-07-21
+
 ### Fixed
 
 - Fixed CVE-2017-7512 [PR #393](https://github.com/3scale/apicast/pull/392)
@@ -172,7 +174,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - Major rewrite using JSON configuration instead of code generation.
 
-[Unreleased]: https://github.com/3scale/apicast/compare/v3.1.0-alpha1...HEAD
+[Unreleased]: https://github.com/3scale/apicast/compare/v3.1.0-beta1...HEAD
 [2.0.0]: https://github.com/3scale/apicast/compare/v0.2...v2.0.0
 [3.0.0-alpha1]: https://github.com/3scale/apicast/compare/v2.0.0...v3.0.0-alpha1
 [3.0.0-alpha2]: https://github.com/3scale/apicast/compare/v3.0.0-alpha1...v3.0.0-alpha2
@@ -181,3 +183,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 [3.0.0-beta3]: https://github.com/3scale/apicast/compare/v3.0.0-beta2...v3.0.0-beta3
 [3.0.0]: https://github.com/3scale/apicast/compare/v3.0.0-beta3...v3.0.0
 [3.1.0-alpha1]: https://github.com/3scale/apicast/compare/v3.0.0...v3.1.0-alpha1
+[3.1.0-beta1]: https://github.com/3scale/apicast/compare/v3.1.0-alpha1...v3.1.0-beta1